# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  entity.CreateUpdateNotePayload:
    properties:
      content:
        example: 1. Nothing.
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: To Do list
        maxLength: 255
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  utils.Response:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.ResponseMeta'
      validation_errors:
        items:
          type: string
        type: array
    type: object
  utils.ResponseMeta:
    properties:
      count:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
host: localhost:9000
info:
  contact: {}
  description: This is a simple notes api server
  title: Simple Notes API
  version: "1.0"
paths:
  /notes:
    get:
      description: Get list of note.
      parameters:
      - description: Pagination page number (default 1, max 500)
        example: 1
        in: query
        name: page
        type: integer
      - description: Pagination data limit  (default 10, max 100)
        example: 10
        in: query
        name: count
        type: integer
      - description: 'Data sorting (available fields: id, title, created_at, updated_at).
          For descending order use prefix '
        example: -created_at
        in: query
        name: sort
        type: string
      - description: Keyword for searching note by title or content
        example: to do list
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get list of note.
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Create new Note.
      parameters:
      - description: note data
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/entity.CreateUpdateNotePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create Note.
      tags:
      - Notes
  /notes/{note_id}:
    delete:
      description: Delete note.
      parameters:
      - description: Note ID
        example: 01HQSH92SNYQVCBDSD38XNBRYM
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete note.
      tags:
      - Notes
    get:
      description: Get note by note ID.
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get note by note ID.
      tags:
      - Notes
    put:
      consumes:
      - application/json
      description: Update existing Note.
      parameters:
      - description: Note ID
        example: 01HQSH92SNYQVCBDSD38XNBRYM
        in: path
        name: note_id
        required: true
        type: string
      - description: note data
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/entity.CreateUpdateNotePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update Note.
      tags:
      - Notes
swagger: "2.0"